Switch things to typescript, use @types/modernizr, fix global.d.ts. Move mobile_guide redirect to index.ts
